+++ .github/workflows/_all_ci.yml
@@ -52,8 +52,8 @@ jobs:
       working-directory: ${{ inputs.working-directory }}
     secrets: inherit
 
-  pydantic-compatibility:
-    uses: ./.github/workflows/_pydantic_compatibility.yml
+  dependencies:
+    uses: ./.github/workflows/_dependencies.yml
     with:
       working-directory: ${{ inputs.working-directory }}
     secrets: inherit
